refactor(web): move the voice output picker into the voice-model entity
It was inside the workflow editor's internals, where the voice agent form
cannot import it. A second copy of a data-driven cascade is worse than a move.
