+++ assemblix-app-web/src/entities/voice-model/index.ts
@@ -5,8 +5,11 @@ export {
   useGetSystemVoicesQuery,
 } from "./api/voice-model.api";
 
+export { VoiceOutputPicker } from "./ui/voice-output-picker";
+
 export type {
   VoiceCapability,
+  VoiceOutputConfig,
   VoiceListItem,
   VoiceModelMetadata,
   VoiceProviderListItem,
