+++ assemblix-app-web/src/entities/workflow/model/types.ts
@@ -1,3 +1,6 @@
+import type { VoiceOutputConfig } from "@/entities/voice-model";
+
+export type { VoiceOutputConfig };
 export interface WorkflowVersion {
   id: string;
   version: number;
@@ -172,18 +175,6 @@ export interface SetVariableNodeConfig extends Record<string, unknown> {
 export type OutputMode = "last_agent" | "specific_agent" | "custom";
 export type FilterMode = "all" | "none" | "selected";
 
-// TTS provider/voice for the END node. `provider` is a voice-provider id
-// (e.g. "elevenlabs"), intentionally a plain string and not the `Provider`
-// enum (which only covers LLM providers).
-export interface VoiceOutputConfig {
-  provider: string;
-  model: string;
-  voiceId?: string;
-  credentialId?: string;
-  // Explicit opt-in to live WS streaming. Only offered for providers that
-  // expose a realtime route (e.g. ElevenLabs); ignored/absent otherwise.
-  realtime?: boolean;
-}
 
 export interface WorkflowAvatarConfig {
   provider: string;
